+++ packages/@justweb3/widget/src/lib/dialogs/ConfigurationDialog/index.tsx
@@ -19,12 +19,14 @@ export interface ConfigurationDialogProps {
   open: boolean;
   handleOpenDialog: (open: boolean) => void;
   logout?: () => void;
+  onBack: () => void
 }
 
 export const ConfigurationDialog: FC<ConfigurationDialogProps> = ({
   open,
   handleOpenDialog,
   logout,
+  onBack
 }) => {
   const {
     config: { disableOverlay },
@@ -72,7 +74,7 @@ export const ConfigurationDialog: FC<ConfigurationDialogProps> = ({
     >
       <Flex justify="space-between" direction="column" gap="20px">
         <Flex direction='row' gap={'10px'} align={'center'} style={{ marginTop: 9 }} >
-          <BackBtn onClick={() => handleOpenDialog(false)} />
+          <BackBtn onClick={() => { handleOpenDialog(false); onBack() }} />
           <H2 style={{
             fontWeight: 700,
           }}>Configure</H2>