+++ packages/@justweb3/widget/src/lib/dialogs/SignInDialog/index.tsx
@@ -1,6 +1,7 @@
 import {
   Records,
   useAccountEnsNames,
+  useAccountInvitations,
   useAccountSubnames,
   useAddSubname,
   useEnsSignIn,
@@ -26,6 +27,7 @@ import {
 import clsx from 'clsx';
 import React, { FC, Fragment, useMemo, useState } from 'react';
 import { SelectSubnameItem } from '../../components/SelectSubnameItem';
+import { SubnameInvitationItem } from '../../components/SubnameInvitationItem';
 import { useDebounce } from '../../hooks/useDebounce';
 import { DefaultDialog } from '../DefaultDialog';
 import styles from './SignInDialog.module.css';
@@ -81,13 +83,19 @@ export const SignInDialog: FC<SignInDialogProps> = ({
   const { isConnected, address } = useMountedAccount();
   const { accountSubnames, isAccountSubnamesPending } = useAccountSubnames();
   const { accountEnsNames, isAccountEnsNamesPending } = useAccountEnsNames();
+  const { invitations, isInvitationsPending, refetchInvitations } =
+    useAccountInvitations({
+      chainId,
+      enabled: !!chainId && !!address,
+    });
 
-  const { primaryName, isPrimaryNameLoading, refetchPrimaryName } = usePrimaryName({
-    address,
-    chainId,
-    enabled: !!address && !!chainId,
-    priority: 'onChain'
-  })
+  const { primaryName, isPrimaryNameLoading, refetchPrimaryName } =
+    usePrimaryName({
+      address,
+      chainId,
+      enabled: !!address && !!chainId,
+      priority: 'onChain',
+    });
 
   const [username, setUsername] = useState('');
   const { debouncedValue: debouncedUsername, isDebouncing } = useDebounce(
@@ -234,16 +242,16 @@ export const SignInDialog: FC<SignInDialogProps> = ({
   ]);
 
   const primarySubname = useMemo(() => {
-    return subnames.find((subname) => subname.ens === primaryName)
-  }, [subnames, primaryName,])
+    return subnames.find((subname) => subname.ens === primaryName);
+  }, [subnames, primaryName]);
 
   const filteredSubnames = useMemo(() => {
-    return subnames.filter((subname) => subname.ens !== primaryName)
-  }, [subnames, primaryName])
+    return subnames.filter((subname) => subname.ens !== primaryName);
+  }, [subnames, primaryName]);
 
   const shouldBeAbleToSelect = useMemo(() => {
-    return subnames.length > 0;
-  }, [subnames]);
+    return subnames.length > 0 || invitations.length > 0;
+  }, [invitations.length, subnames.length]);
 
   const shouldBeAbleToClaim = useMemo(() => {
     return !subnames.find((subname) => subname.ens.endsWith(claimableEns));
@@ -286,25 +294,35 @@ export const SignInDialog: FC<SignInDialogProps> = ({
           </SPAN>
         </Badge>
         {isAccountSubnamesPending ||
-          isAccountEnsNamesPending ||
-          isOffchainResolversPending ? (
+        isInvitationsPending ||
+        isAccountEnsNamesPending ||
+        isOffchainResolversPending ? (
           <div className={styles.loadingContainer}>
             <LoadingSpinner color={'var(--justweb3-primary-color)'} />
           </div>
         ) : (
-          <Flex direction="column" gap="20px">
+          <Flex direction="column" gap="10px">
             <TransitionElement
               visible={shouldBeAbleToSelect}
               maxheight="fit-content"
             >
-              <Flex direction="column" gap="20px" justify={'space-between'}>
+              <Flex direction="column" gap="10px" justify={'space-between'}>
                 <H2>Select an ENS</H2>
                 <Flex
                   direction="column"
                   gap="15px"
                   className={clsx(styles.contentWrapper)}
                 >
-                  {!isPrimaryNameLoading && primarySubname &&
+                  {invitations.length > 0 &&
+                    invitations.map((inv, index) => (
+                      <Fragment key={'subname-invitation-' + index}>
+                        <SubnameInvitationItem
+                          subname={inv}
+                          onInvitationChange={refetchInvitations}
+                        />
+                      </Fragment>
+                    ))}
+                  {!isPrimaryNameLoading && primarySubname && (
                     <SelectSubnameItem
                       selectedSubname={subnameSigningIn}
                       subname={primarySubname}
@@ -318,7 +336,7 @@ export const SignInDialog: FC<SignInDialogProps> = ({
                       }}
                       isPrimary={true}
                     />
-                  }
+                  )}
                   {filteredSubnames.map((subname, index) => (
                     <Fragment key={'subname-' + index}>
                       <SelectSubnameItem