+++ packages/@justweb3/xmtp-plugin/src/stories/xmtp.stories.tsx
@@ -1,4 +1,5 @@
 import {
+  JustEnsCard,
   JustWeb3Button,
   JustWeb3Provider,
   JustWeb3ProviderConfig,