+++ packages/@justaname.id/sdk/CHANGELOG.md
@@ -1,3 +1,108 @@
+## 0.2.174 (2025-02-06)
+
+
+### 🧱 Updated Dependencies
+
+- Updated @justaname.id/siwens to 0.0.107
+
+## 0.2.173 (2025-02-06)
+
+
+### 🧱 Updated Dependencies
+
+- Updated @justaname.id/siwens to 0.0.106
+
+## 0.2.172 (2025-02-06)
+
+
+### 🧱 Updated Dependencies
+
+- Updated @justaname.id/siwens to 0.0.105
+
+## 0.2.171 (2025-02-03)
+
+
+### 🧱 Updated Dependencies
+
+- Updated @justaname.id/siwens to 0.0.104
+
+## 0.2.170 (2025-01-29)
+
+
+### 🧱 Updated Dependencies
+
+- Updated @justaname.id/siwens to 0.0.103
+
+## 0.2.169 (2025-01-28)
+
+
+### 🧱 Updated Dependencies
+
+- Updated @justaname.id/siwens to 0.0.102
+
+## 0.2.168 (2025-01-28)
+
+
+### 🧱 Updated Dependencies
+
+- Updated @justaname.id/siwens to 0.0.101
+
+## 0.2.167 (2025-01-28)
+
+
+### 🧱 Updated Dependencies
+
+- Updated @justaname.id/siwens to 0.0.100
+
+## 0.2.166 (2025-01-22)
+
+
+### 🧱 Updated Dependencies
+
+- Updated @justaname.id/siwens to 0.0.99
+
+## 0.2.165 (2025-01-16)
+
+
+### 🧱 Updated Dependencies
+
+- Updated @justaname.id/siwens to 0.0.98
+
+## 0.2.164 (2025-01-16)
+
+
+### 🧱 Updated Dependencies
+
+- Updated @justaname.id/siwens to 0.0.97
+
+## 0.2.163 (2025-01-16)
+
+
+### 🧱 Updated Dependencies
+
+- Updated @justaname.id/siwens to 0.0.96
+
+## 0.2.162 (2024-12-22)
+
+
+### 🧱 Updated Dependencies
+
+- Updated @justaname.id/siwens to 0.0.95
+
+## 0.2.161 (2024-12-17)
+
+
+### 🧱 Updated Dependencies
+
+- Updated @justaname.id/siwens to 0.0.94
+
+## 0.2.160 (2024-12-17)
+
+
+### 🧱 Updated Dependencies
+
+- Updated @justaname.id/siwens to 0.0.93
+
 ## 0.2.159 (2024-12-12)