+++ packages/@justverified/plugin/CHANGELOG.md
@@ -1,3 +1,71 @@
+## 0.0.112 (2025-02-06)
+
+This was a version bump only for @justverified/plugin to align it with other projects, there were no code changes.
+
+## 0.0.111 (2025-02-06)
+
+This was a version bump only for @justverified/plugin to align it with other projects, there were no code changes.
+
+## 0.0.110 (2025-02-06)
+
+
+### 🩹 Fixes
+
+- **justVerified:** typo in expiry date ([638347ca](https://github.com/JustaName-id/JustaName-sdk/commit/638347ca))
+
+
+### ❤️  Thank You
+
+- anthony2399 @anthony23991
+
+## 0.0.109 (2025-02-03)
+
+This was a version bump only for @justverified/plugin to align it with other projects, there were no code changes.
+
+## 0.0.108 (2025-01-29)
+
+This was a version bump only for @justverified/plugin to align it with other projects, there were no code changes.
+
+## 0.0.107 (2025-01-28)
+
+This was a version bump only for @justverified/plugin to align it with other projects, there were no code changes.
+
+## 0.0.106 (2025-01-28)
+
+This was a version bump only for @justverified/plugin to align it with other projects, there were no code changes.
+
+## 0.0.105 (2025-01-28)
+
+This was a version bump only for @justverified/plugin to align it with other projects, there were no code changes.
+
+## 0.0.104 (2025-01-22)
+
+This was a version bump only for @justverified/plugin to align it with other projects, there were no code changes.
+
+## 0.0.103 (2025-01-16)
+
+This was a version bump only for @justverified/plugin to align it with other projects, there were no code changes.
+
+## 0.0.102 (2025-01-16)
+
+This was a version bump only for @justverified/plugin to align it with other projects, there were no code changes.
+
+## 0.0.101 (2025-01-16)
+
+This was a version bump only for @justverified/plugin to align it with other projects, there were no code changes.
+
+## 0.0.100 (2024-12-22)
+
+This was a version bump only for @justverified/plugin to align it with other projects, there were no code changes.
+
+## 0.0.99 (2024-12-17)
+
+This was a version bump only for @justverified/plugin to align it with other projects, there were no code changes.
+
+## 0.0.98 (2024-12-17)
+
+This was a version bump only for @justverified/plugin to align it with other projects, there were no code changes.
+
 ## 0.0.97 (2024-12-12)
 
 This was a version bump only for @justverified/plugin to align it with other projects, there were no code changes.