+++ packages/siwens/CHANGELOG.md
@@ -1,3 +1,51 @@
+## 0.1.14 (2025-02-06)
+
+This was a version bump only for siwens to align it with other projects, there were no code changes.
+
+## 0.1.13 (2025-02-06)
+
+This was a version bump only for siwens to align it with other projects, there were no code changes.
+
+## 0.1.12 (2025-02-06)
+
+This was a version bump only for siwens to align it with other projects, there were no code changes.
+
+## 0.1.11 (2025-02-03)
+
+This was a version bump only for siwens to align it with other projects, there were no code changes.
+
+## 0.1.10 (2025-01-29)
+
+This was a version bump only for siwens to align it with other projects, there were no code changes.
+
+## 0.1.9 (2025-01-28)
+
+This was a version bump only for siwens to align it with other projects, there were no code changes.
+
+## 0.1.8 (2025-01-28)
+
+This was a version bump only for siwens to align it with other projects, there were no code changes.
+
+## 0.1.7 (2025-01-28)
+
+This was a version bump only for siwens to align it with other projects, there were no code changes.
+
+## 0.1.6 (2025-01-22)
+
+This was a version bump only for siwens to align it with other projects, there were no code changes.
+
+## 0.1.5 (2025-01-16)
+
+This was a version bump only for siwens to align it with other projects, there were no code changes.
+
+## 0.1.4 (2025-01-16)
+
+This was a version bump only for siwens to align it with other projects, there were no code changes.
+
+## 0.1.3 (2025-01-16)
+
+This was a version bump only for siwens to align it with other projects, there were no code changes.
+
 ## 0.1.2 (2024-12-11)
 
 This was a version bump only for siwens to align it with other projects, there were no code changes.